[fix] #349 seperate cards collection in paper, change hide timer balloon logic
